Stop publishing any of the zip symbols as extern
When compiling both SSZipArchive and SSOldZipArchive into the same module, I ran into issues where the extern symbols from the zip methods collided and could not be linked.

The solution came in recognizing that we don't need any of those symbols to be published, because SSOldZipArchive wraps all uses of those minizip symbols.

This commit hides all the minizip symbols that were causing linking issues.